fix(cdn): read manifest file sizes as 64-bit
Bumps ritoshark to c6f2d39, where the manifest file-size field is read at its real
eight-byte width. It was read as four bytes, which returns the correct value for
every file below 4 GiB and silently wraps above - no error, just a wrong number
that the download planner would have trusted.

Nothing had failed because League's largest shipped file is 3.11 GiB, inside the
range where both widths agree. The FlatBuffer vtable is what settles it: the
eight-byte fields pack together at 16, 24, 32, 40, so size at 32 owns eight bytes.

FilePlan::size was already u64, so the cast at the boundary is now redundant and
drops out.

Hematite moves to f63c656 in step. Its rs_* pins must equal the workspace
ritoshark rev or the lock carries two copies of rs_io/rs_bin/rs_wad/rs_tex/rs_mesh;
verified the lock holds exactly one of each. That rev changes nothing for Hematite -
c6f2d39 touches only rs_rman, which it does not depend on.

+++ src-tauri/crates/flint-net/src/cdn/downloader.rs
@@ -92,7 +92,7 @@ pub fn plan_download(manifest: &crate::cdn::manifest::Manifest, indices: &[usize
             let entry = manifest.file(index)?;
             Some(FilePlan {
                 rel_path: paths.get(index).map(|(p, _)| p.clone()).unwrap_or_default(),
-                size: entry.size as u64,
+                size: entry.size,
                 chunks: manifest.file_chunks(index),
                 hash_type: manifest.rman.file_hash_type(entry),
             })
